Plane: Fix throttle nudge in GUIDED mode
guided_state.target_airspeed_cm is initialised to -1 and when OFFBOARD_GUIDED == ENABLED, check for is_zero was not correct, causing throttle nudge not working in GUIDED mode.
Also > 0.0 changed to is_positive, to keep in line with other code
Testing script is updated to put throttle stick to neutral where GUIDED mode is used.